Actually remove LongDoubleBits
