+++ libc/src/__support/FPUtil/x86_64/NextAfterLongDouble.h
@@ -61,7 +61,7 @@ LIBC_INLINE long double nextafter(long double from, long double to) {
         from_bits.set_biased_exponent(from_bits.get_biased_exponent() + 1);
         if (from_bits.is_inf())
           raise_except_if_required(FE_OVERFLOW | FE_INEXACT);
-        return from_bits;
+        return from_bits.get_val();;
       } else {
         ++int_val;
       }
@@ -75,7 +75,7 @@ LIBC_INLINE long double nextafter(long double from, long double to) {
         // from == 0 is handled separately so decrementing the exponent will not
         // lead to underflow.
         from_bits.set_biased_exponent(from_bits.get_biased_exponent() - 1);
-        return from_bits;
+        return from_bits.get_val();;
       } else {
         --int_val;
       }
@@ -94,7 +94,7 @@ LIBC_INLINE long double nextafter(long double from, long double to) {
         // from == 0 is handled separately so decrementing the exponent will not
         // lead to underflow.
         from_bits.set_biased_exponent(from_bits.get_biased_exponent() - 1);
-        return from_bits;
+        return from_bits.get_val();;
       } else {
         --int_val;
       }
@@ -109,7 +109,7 @@ LIBC_INLINE long double nextafter(long double from, long double to) {
         from_bits.set_biased_exponent(from_bits.get_biased_exponent() + 1);
         if (from_bits.is_inf())
           raise_except_if_required(FE_OVERFLOW | FE_INEXACT);
-        return from_bits;
+        return from_bits.get_val();;
       } else {
         ++int_val;
       }
